+++ src/main/scala/com/github/Chronox/pool/db/Share.scala
@@ -1,12 +1,13 @@
 package com.github.Chronox.pool.db
+import java.lang.Long
 import java.math.BigInteger
 
 class Share(
   var userId: Long,
   var blockId: BigInteger,
   var nonce: Long,
-  var deadline: BigInteger,
+  var deadline: Long,
   var isPaid: Boolean
-  ) {
-  def this() = this(0, BigInteger.valueOf(0), 0, BigInteger.valueOf(0), false)
+) {
+  def this() = this(0, BigInteger.valueOf(0), 0, 0L, false)
 }
